requirements.txt
opencv-python
mediapipe
numpy
pillow
tkinter